// packages/ocom/graphql/src/schema/builder/generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  AccountNumber: { input: any; output: any };
  BigInt: { input: any; output: any };
  Byte: { input: any; output: any };
  CountryCode: { input: any; output: any };
  CountryName: { input: any; output: any };
  Cuid: { input: any; output: any };
  Currency: { input: any; output: any };
  DID: { input: any; output: any };
  Date: { input: Date; output: Date };
  DateTime: { input: any; output: any };
  DateTimeISO: { input: any; output: any };
  DeweyDecimal: { input: any; output: any };
  Duration: { input: any; output: any };
  EmailAddress: { input: string; output: string };
  GUID: { input: string; output: string };
  GeoJSON: { input: any; output: any };
  HSL: { input: any; output: any };
  HSLA: { input: any; output: any };
  HexColorCode: { input: any; output: any };
  Hexadecimal: { input: any; output: any };
  IBAN: { input: any; output: any };
  IP: { input: any; output: any };
  IPCPatent: { input: any; output: any };
  IPv4: { input: any; output: any };
  IPv6: { input: any; output: any };
  ISBN: { input: any; output: any };
  ISO8601Duration: { input: any; output: any };
  JSON: { input: any; output: any };
  JSONObject: { input: any; output: any };
  JWT: { input: any; output: any };
  LCCSubclass: { input: any; output: any };
  Latitude: { input: any; output: any };
  LocalDate: { input: any; output: any };
  LocalDateTime: { input: any; output: any };
  LocalEndTime: { input: any; output: any };
  LocalTime: { input: any; output: any };
  Locale: { input: any; output: any };
  Long: { input: any; output: any };
  Longitude: { input: any; output: any };
  MAC: { input: any; output: any };
  NegativeFloat: { input: any; output: any };
  NegativeInt: { input: any; output: any };
  NonEmptyString: { input: any; output: any };
  NonNegativeFloat: { input: any; output: any };
  NonNegativeInt: { input: any; output: any };
  NonPositiveFloat: { input: any; output: any };
  NonPositiveInt: { input: any; output: any };
  ObjectID: { input: any; output: any };
  PhoneNumber: { input: any; output: any };
  Port: { input: any; output: any };
  PositiveFloat: { input: any; output: any };
  PositiveInt: { input: any; output: any };
  PostalCode: { input: any; output: any };
  RGB: { input: any; output: any };
  RGBA: { input: any; output: any };
  RoutingNumber: { input: any; output: any };
  SESSN: { input: any; output: any };
  SafeInt: { input: any; output: any };
  SemVer: { input: any; output: any };
  Time: { input: any; output: any };
  TimeZone: { input: any; output: any };
  Timestamp: { input: any; output: any };
  URL: { input: any; output: any };
  USCurrency: { input: any; output: any };
  UUID: { input: any; output: any };
  UnsignedFloat: { input: any; output: any };
  UnsignedInt: { input: any; output: any };
  UtcOffset: { input: any; output: any };
  Void: { input: any; output: any };
};

export type BlobAuthHeader = {
  readonly __typename?: "BlobAuthHeader";
  readonly authHeader?: Maybe<Scalars["String"]["output"]>;
  readonly blobName?: Maybe<Scalars["String"]["output"]>;
  readonly blobPath?: Maybe<Scalars["String"]["output"]>;
  readonly indexTags?: Maybe<ReadonlyArray<Maybe<BlobIndexTag>>>;
  readonly metadataFields?: Maybe<ReadonlyArray<Maybe<BlobMetadataField>>>;
  readonly requestDate?: Maybe<Scalars["String"]["output"]>;
};

export type BlobIndexTag = {
  readonly __typename?: "BlobIndexTag";
  readonly name: Scalars["String"]["output"];
  readonly value: Scalars["String"]["output"];
};

export type BlobMetadataField = {
  readonly __typename?: "BlobMetadataField";
  readonly name: Scalars["String"]["output"];
  readonly value: Scalars["String"]["output"];
};

/**  Required to enable Apollo Cache Control  */
export type CacheControlScope = "PRIVATE" | "PUBLIC";

export type Community = MongoBase & {
  readonly __typename?: "Community";
  readonly createdAt: Scalars["DateTime"]["output"];
  readonly createdBy: EndUser;
  readonly domain?: Maybe<Scalars["String"]["output"]>;
  readonly handle?: Maybe<Scalars["String"]["output"]>;
  readonly id: Scalars["ObjectID"]["output"];
  readonly name: Scalars["String"]["output"];
  readonly publicContentBlobUrl?: Maybe<Scalars["String"]["output"]>;
  readonly schemaVersion: Scalars["String"]["output"];
  readonly updatedAt: Scalars["DateTime"]["output"];
  readonly whiteLabelDomain?: Maybe<Scalars["String"]["output"]>;
};

export type CommunityCreateInput = {
  readonly name: Scalars["String"]["input"];
};

export type CommunityMutationResult = MutationResult & {
  readonly __typename?: "CommunityMutationResult";
  readonly community?: Maybe<Community>;
  readonly status: MutationStatus;
};

export type EndUser = MongoBase & {
  readonly __typename?: "EndUser";
  readonly accessBlocked?: Maybe<Scalars["Boolean"]["output"]>;
  readonly createdAt?: Maybe<Scalars["DateTime"]["output"]>;
  readonly displayName?: Maybe<Scalars["String"]["output"]>;
  readonly externalId?: Maybe<Scalars["String"]["output"]>;
  readonly id: Scalars["ObjectID"]["output"];
  readonly personalInformation?: Maybe<EndUserPersonalInformation>;
  readonly schemaVersion?: Maybe<Scalars["String"]["output"]>;
  readonly tags?: Maybe<ReadonlyArray<Maybe<Scalars["String"]["output"]>>>;
  readonly updatedAt?: Maybe<Scalars["DateTime"]["output"]>;
};
// ...
export type MutationResult = {
  readonly status: MutationStatus;
};
// ...
export type ResolversObject<TObject> = WithIndex<TObject>;
// ...
export type BlobAuthHeaderResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["BlobAuthHeader"] = ResolversParentTypes["BlobAuthHeader"],
> = ResolversObject<{
  authHeader?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  blobName?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  blobPath?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  indexTags?: Resolver<Maybe<ReadonlyArray<Maybe<ResolversTypes["BlobIndexTag"]>>>, ParentType, ContextType>;
  metadataFields?: Resolver<Maybe<ReadonlyArray<Maybe<ResolversTypes["BlobMetadataField"]>>>, ParentType, ContextType>;
  requestDate?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type BlobIndexTagResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["BlobIndexTag"] = ResolversParentTypes["BlobIndexTag"],
> = ResolversObject<{
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  value?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type BlobMetadataFieldResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["BlobMetadataField"] = ResolversParentTypes["BlobMetadataField"],
> = ResolversObject<{
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  value?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export interface ByteScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes["Byte"], any> {
  name: "Byte";
}

export type CommunityResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["Community"] = ResolversParentTypes["Community"],
> = ResolversObject<{
  createdAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  createdBy?: Resolver<ResolversTypes["EndUser"], ParentType, ContextType>;
  domain?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  handle?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes["ObjectID"], ParentType, ContextType>;
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  publicContentBlobUrl?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  schemaVersion?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  updatedAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  whiteLabelDomain?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type CommunityMutationResultResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["CommunityMutationResult"] = ResolversParentTypes["CommunityMutationResult"],
> = ResolversObject<{
  community?: Resolver<Maybe<ResolversTypes["Community"]>, ParentType, ContextType>;
  status?: Resolver<ResolversTypes["MutationStatus"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
// ...
export type Resolvers<ContextType = GraphContext> = ResolversObject<{
  AccountNumber?: GraphQLScalarType;
  BigInt?: GraphQLScalarType;
  BlobAuthHeader?: BlobAuthHeaderResolvers<ContextType>;
  BlobIndexTag?: BlobIndexTagResolvers<ContextType>;
  BlobMetadataField?: BlobMetadataFieldResolvers<ContextType>;
  Byte?: GraphQLScalarType;
  Community?: CommunityResolvers<ContextType>;
  CommunityMutationResult?: CommunityMutationResultResolvers<ContextType>;
  CountryCode?: GraphQLScalarType;
  CountryName?: GraphQLScalarType;
  Cuid?: GraphQLScalarType;
  Currency?: GraphQLScalarType;
  DID?: GraphQLScalarType;
  Date?: GraphQLScalarType;
  DateTime?: GraphQLScalarType;
  DateTimeISO?: GraphQLScalarType;
  DeweyDecimal?: GraphQLScalarType;
  Duration?: GraphQLScalarType;
  EmailAddress?: GraphQLScalarType;
  EndUser?: EndUserResolvers<ContextType>;
  EndUserContactInformation?: EndUserContactInformationResolvers<ContextType>;
  EndUserIdentityDetails?: EndUserIdentityDetailsResolvers<ContextType>;
  EndUserPersonalInformation?: EndUserPersonalInformationResolvers<ContextType>;
  GUID?: GraphQLScalarType;
  GeoJSON?: GraphQLScalarType;
  HSL?: GraphQLScalarType;
  HSLA?: GraphQLScalarType;
  HexColorCode?: GraphQLScalarType;
  Hexadecimal?: GraphQLScalarType;
  IBAN?: GraphQLScalarType;
  IP?: GraphQLScalarType;
  IPCPatent?: GraphQLScalarType;
  IPv4?: GraphQLScalarType;
  IPv6?: GraphQLScalarType;
  ISBN?: GraphQLScalarType;
  ISO8601Duration?: GraphQLScalarType;
  JSON?: GraphQLScalarType;
  JSONObject?: GraphQLScalarType;
  JWT?: GraphQLScalarType;
  LCCSubclass?: GraphQLScalarType;
  Latitude?: GraphQLScalarType;
  LocalDate?: GraphQLScalarType;
  LocalDateTime?: GraphQLScalarType;
  LocalEndTime?: GraphQLScalarType;
  LocalTime?: GraphQLScalarType;
  Locale?: GraphQLScalarType;
  Long?: GraphQLScalarType;
  Longitude?: GraphQLScalarType;
  MAC?: GraphQLScalarType;
  Member?: MemberResolvers<ContextType>;
  MemberAccount?: MemberAccountResolvers<ContextType>;
  MemberProfile?: MemberProfileResolvers<ContextType>;
  MongoBase?: MongoBaseResolvers<ContextType>;
  MongoSubdocument?: MongoSubdocumentResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
  MutationResult?: MutationResultResolvers<ContextType>;
  MutationStatus?: MutationStatusResolvers<ContextType>;
  NegativeFloat?: GraphQLScalarType;
  NegativeInt?: GraphQLScalarType;
  NonEmptyString?: GraphQLScalarType;
  NonNegativeFloat?: GraphQLScalarType;
  NonNegativeInt?: GraphQLScalarType;
  NonPositiveFloat?: GraphQLScalarType;
  NonPositiveInt?: GraphQLScalarType;
  ObjectID?: GraphQLScalarType;
  PhoneNumber?: GraphQLScalarType;
  Port?: GraphQLScalarType;
  PositiveFloat?: GraphQLScalarType;
  PositiveInt?: GraphQLScalarType;
  PostalCode?: GraphQLScalarType;
  Query?: QueryResolvers<ContextType>;
  RGB?: GraphQLScalarType;
  RGBA?: GraphQLScalarType;
  RoutingNumber?: GraphQLScalarType;
  SESSN?: GraphQLScalarType;
  SafeInt?: GraphQLScalarType;
  SemVer?: GraphQLScalarType;
  Time?: GraphQLScalarType;
  TimeZone?: GraphQLScalarType;
  Timestamp?: GraphQLScalarType;
  URL?: GraphQLScalarType;
  USCurrency?: GraphQLScalarType;
  UUID?: GraphQLScalarType;
  UnsignedFloat?: GraphQLScalarType;
  UnsignedInt?: GraphQLScalarType;
  UtcOffset?: GraphQLScalarType;
  Void?: GraphQLScalarType;
}>;